refactor(ai-agent): split skill_manager into submodules (Stage 9.2, M5)
Purely mechanical split — function bodies byte-identical.

- skill_manager/mod.rs:  struct + AcpSkillManager impl + detect_skill_load_request
                         + extract_body + discovery-side tests
- skill_manager/prompt_builder.rs:  pure prompt/instruction builders + their tests

Public API unchanged — prompt_builder re-exported via `pub use prompt_builder::*`.
Each sub-file ≤ 400 lines. No behavior change.
